# v1.1/backend/docs/swagger.yaml
definitions:
...
  common.DeleteRequest:
    properties:
      id:
        example: ""
        type: string
    required:
    - id
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  picture.ListPictureResponse:
    properties:
      current:
        description: 当前页数
        type: integer
      pages:
        description: 总页数
        type: integer
      records:
        items:
          $ref: '#/definitions/entity.Picture'
        type: array
      size:
        description: 页面大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
...
  picture.PictureEditByBatchRequest:
    properties:
      category:
        description: 分类
        type: string
      nameRule:
        description: 名称规则，暂时只支持“名称{序号}的形式，序号将会自动递增”
        type: string
      pictureIdList:
        description: 图片ID列表
        items:
          type: string
        type: array
      spaceId:
        description: 空间ID
        example: ""
        type: string
      tags:
        description: 标签
        items:
          type: string
        type: array
    type: object
  picture.PictureEditRequest:
    properties:
      category:
        type: string
      id:
        example: ""
        type: string
      introduction:
        type: string
      name:
        type: string
      spaceId:
        description: 空间ID
        example: ""
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8001
info:
  contact:
    email: lwh24621@qq.com
    name: lwhhhh
    url: 无
  description: CanvasCloud
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: 无服务条款
  title: CanvasCloud
  version: "1.0"
paths:
...
  /v1/picture/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片的ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/common.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 删除失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据ID软删除图片「登录校验」
      tags:
      - picture
  /v1/picture/edit:
    post:
      consumes:
      - application/json
      description: 若图片不存在，则返回false
      parameters:
      - description: 需要更新的图片信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 编辑图片
      tags:
      - picture
  /v1/picture/edit/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量的图片ID、空间ID和分类和标签
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureEditByBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 批量更新图片请求「登录校验」
      tags:
      - picture
  /v1/picture/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 图片的ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Picture'
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据ID获取图片「管理员」
      tags:
      - picture
  /v1/picture/get/vo:
    get:
      consumes:
      - application/json
      parameters:
      - description: 图片的ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.PictureVO'
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据ID获取脱敏的图片
      tags:
      - picture
  /v1/picture/list/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 需要查询的页数、以及图片关键信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.ListPictureResponse'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 分页获取一系列图片信息「管理员」
      tags:
      - picture
  /v1/picture/list/page/vo/cache:
    post:
      consumes:
      - application/json
      parameters:
      - description: 需要查询的页数、以及图片关键信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.ListPictureVOResponse'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 带有缓存的分页获取一系列图片信息
      tags:
      - picture
  /v1/picture/list/page/vo/procache:
    post:
      consumes:
      - application/json
      parameters:
      - description: 需要查询的页数、以及图片关键信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.ListPictureVOResponse'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 带有缓存的分页获取一系列图片信息
      tags:
      - picture
  /v1/picture/out_painting/create_task:
    get:
      consumes:
      - application/json
      parameters:
      - description: 任务的ID
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回任务进展信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.GetOutPaintingResponse'
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取AI扩图任务信息「登录校验」
      tags:
      - picture
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建扩图任务所需信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.CreateOutPaintingTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回任务信息
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.CreateOutPaintingTaskResponse'
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建AI扩图任务请求「登录校验」
      tags:
      - picture
...
  /v1/picture/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审核图片所需信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 审核更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 执行图片审核「管理员」
      tags:
      - picture
  /v1/picture/search/color:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片的颜色和空间ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureSearchByColorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/picture.PictureVO'
                  type: array
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据图片的颜色搜索相似图片「登录校验」
      tags:
      - picture
  /v1/picture/search/picture:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片的ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/picture.PictureSearchByPictureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ImageSearchResult'
                  type: array
              type: object
        "400":
          description: 获取失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 根据图片ID搜索图片
      tags:
      - picture
  /v1/picture/tag_category:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/picture.PictureTagCategory'
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取图片的标签和分类（固定）
      tags:
      - picture
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"